requirements.txt
langchain==0.2.13
langchain-community==0.2.12
langchain-google-genai==1.0.9
flask==3.0.3
flask-cors==5.0.0
chromadb==0.5.18
python-dotenv==1.0.1
tiktoken==0.7.0
sentence-transformers==3.0.1